requirements.txt
fastapi
uvicorn
sqlalchemy
#python -m pip install "pymongo[srv]==3.11"
# 'pymongo[srv]==3.11'
python-dotenv
psycopg2-binary   #This is used to connect to PostgreSQL

#Day_5
pymysql
#urllib

#Day_6

#Day_7
jinja2

#Day_8
mangum

#Day_9
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1

#Day_10
pydantic
pydantic[email]

#Day_11
redis
httpx

#Day_12
websockets

#Day_14
httpx
pytest

#Day_15
sqlmodel
aiosqliteA

#Day_20
fastapi[standard]

#Day_22
sqlmodel

#Day_23
strawberry-graphql

#Day_24
poetry

#Day_25
opentelemetry-distro

#Day_26
supabase
jinja2
python-multipart
email-validator
python-dotenv
pyjwt